requirements.txt
tensorflow-gpu
pillow
lxml
jupyter
matplotlib
flask

